paper-version/requirements.txt
# Python 3.11
aiohttp==3.11.10
filelock==3.15.4
numpy==1.24.4
pyyaml==6.0.1
trafilatura==2.0.0
langchain==0.3.15
langchain-community==0.3.15
langchain-huggingface==0.1.2
openai==1.86.0
sentence-transformers==2.7.0
faiss-cpu==1.8.0.post1